Updating scripts
Created a new backup structure for my sensitive files on external drive. Because of this, we had to modify the recovery for i3 install. We have also include setup for external auto mount/auto login.

Adjustments to handle the ssh.tar file and the onwership and mods for the root files